LayoutTests/http/tests/w3c/html/dom/documents/dom-tree-accessors/document.body-getter-expected.txt


PASS Childless document 
PASS Childless html element 
FAIL Body followed by frameset inside the html element assert_equals: expected Element node <body></body> but got Element node <frameset></frameset>
PASS Frameset followed by body inside the html element 
FAIL Body followed by frameset inside a non-HTML html element assert_equals: expected null but got Element node <frameset></frameset>
FAIL Frameset followed by body inside a non-HTML html element assert_equals: expected null but got Element node <frameset></frameset>
PASS Non-HTML body followed by body inside the html element 
PASS Non-HTML frameset followed by body inside the html element 
PASS Body inside an x element followed by a body 
PASS Frameset inside an x element followed by a frameset 
PASS Body as the root node 
PASS Frameset as the root node 
FAIL Body as the root node with a frameset child assert_equals: expected null but got Element node <frameset></frameset>
FAIL Frameset as the root node with a body child assert_equals: expected null but got Element node <body></body>
PASS Non-HTML body as the root node 
PASS Non-HTML frameset as the root node 
